+++ changelog.md
@@ -1,5 +1,6 @@
 
 ## TODO
+	normalize lilyDocument, extract voices.
 	keep comments in LilyDocument
 	lyrics hilighting
 		match chord with voice in the same staff